+++ src/listManagers/UtilsFunctions.js
